Modify 'createList in load' into module
